Free static LinkedRoot when no subscribers
When the last subscriber unsubscribes from a service, free the static
LinkedRoot pointer and set it to NULL. This ensures we don't leak memory
if a service never gets another subscriber. It also doesn't hide true
memory leaks when the app quits; currently each service "leaks" around 8
bytes when the app closes.